@piyushroshan 2024-12-30
Benchmarks: Current
```
go test -bench=. ./internal/corazawaf
goos: linux
goarch: amd64
pkg: github.com/corazawaf/coraza/v3/internal/corazawaf
cpu: xxx   
BenchmarkAddTransformationUnique-24      5506352               354.0 ns/op
BenchmarkAddTransformationSame-24       24820352               167.9 ns/op
BenchmarkTxGetField-24                   3960561               298.3 ns/op           400 B/op          8 allocs/op
BenchmarkTransactionCreation-24            82832             16021 ns/op
BenchmarkTransactionTimestamped-24      13426234                76.43 ns/op
BenchmarkMacro/%{tx.a}-24               34545536                37.93 ns/op
BenchmarkMacro/%{tx.a}_%{tx.b}-24       10972236               111.2 ns/op
BenchmarkMacro/goodbye_world-24         317328151                3.847 ns/op
PASS
ok      github.com/corazawaf/coraza/v3/internal/corazawaf       18.292s
```
---------
Before sync map
```
go test -bench=. ./internal/corazawaf
goos: linux
goarch: amd64
pkg: github.com/corazawaf/coraza/v3/internal/corazawaf
cpu: xxx    
BenchmarkAddTransformationUnique-24       223730             80598 ns/op
BenchmarkAddTransformationSame-24       10833765               121.9 ns/op
BenchmarkTxGetField-24                   3806456               378.3 ns/op           400 B/op          8 allocs/op
BenchmarkTransactionCreation-24            94303             17960 ns/op
BenchmarkTransactionTimestamped-24      15407229                79.22 ns/op
BenchmarkMacro/%{tx.a}-24               29172850                36.42 ns/op
BenchmarkMacro/%{tx.a}_%{tx.b}-24       11480709               119.7 ns/op
BenchmarkMacro/goodbye_world-24         301710660                3.767 ns/op
PASS
ok      github.com/corazawaf/coraza/v3/internal/corazawaf       29.150s
```