@fzipi 2024-12-30
Now I'm getting:

### Current
```
❯ go test -bench=. ./internal/corazawaf
goos: darwin
goarch: amd64
pkg: github.com/corazawaf/coraza/v3/internal/corazawaf
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkAddTransformationUnique-16    	 3311302	       614.1 ns/op
BenchmarkAddTransformationSame-16      	19567063	       633.8 ns/op
BenchmarkTxGetField-16                 	 1000000	      1145 ns/op	     400 B/op	       8 allocs/op
BenchmarkTransactionCreation-16        	   48079	     25621 ns/op
BenchmarkTransactionTimestamped-16     	 4050920	       339.2 ns/op
BenchmarkMacro/%{tx.a}-16              	13499713	        89.01 ns/op
BenchmarkMacro/%{tx.a}_%{tx.b}-16      	 3840274	       341.4 ns/op
BenchmarkMacro/goodbye_world-16        	157292704	         7.671 ns/op
PASS
ok  	github.com/corazawaf/coraza/v3/internal/corazawaf	31.043s
```

### Previous

```
❯ go test -bench=. ./internal/corazawaf
goos: darwin
goarch: amd64
pkg: github.com/corazawaf/coraza/v3/internal/corazawaf
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkTxGetField-16                	 1377608	       868.5 ns/op	     400 B/op	       8 allocs/op
BenchmarkTransactionCreation-16       	   36501	     32750 ns/op
BenchmarkTransactionTimestamped-16    	 4094943	       291.6 ns/op
BenchmarkMacro/%{tx.a}-16             	12492490	        94.61 ns/op
BenchmarkMacro/%{tx.a}_%{tx.b}-16     	 3729914	       318.2 ns/op
BenchmarkMacro/goodbye_world-16       	162130626	         7.388 ns/op
PASS
ok  	github.com/corazawaf/coraza/v3/internal/corazawaf	10.590s
```