@fzipi 2024-12-30
I was lazy and just did checkout the last commit before the merge. Should I add the benchmarks only to the old commits?