@piyushroshan 2024-12-30
Need to copy the two functions to old commit and run. 
Also use before the sync.map merge commit id.